Create CI & CD for develop & TST and Refactor TestDB integration tests
Создание CI/CD для develop и TST, удаление TestDB из CI и рефакторинг тестов
- Разделены CI и CD пайплайны: CI для develop, CD деплоит в TST
- Вынесены общие шаги сборки в .github/actions/build-dotnet/action.yml
- Интеграционные тесты с TestDB перенесены в Api.UnitTests.IntegrationTestDB и исключены из CI
- Улучшена конфигурация: переход на appsettings.json + Key (обновлён ApiConfigurations.cs)
- Добавлена поддержка отчётов о покрытии тестами в CI
