+++ .github/actions/build-dotnet/action.yml
@@ -1,0 +1,30 @@
+name: build-dotnet
+description: ".NET restore + build"
+
+runs:
+    using: composite
+    steps:
+      - name: Setup .NET SDK (from global.json)
+        uses: actions/setup-dotnet@v4
+        with:
+          global-json-file: global.json
+
+      - name: Cache NuGet packages
+        uses: actions/cache@v4
+        with:
+          path: ~/.nuget/packages
+          key: ${{ runner.os }}-nuget-${{ hashFiles('**/*.csproj', 'global.json') }}
+          restore-keys: |
+            ${{ runner.os }}-nuget-
+
+      - name: Clean Release
+        run: dotnet clean --configuration Release
+        shell: bash
+
+      - name: Restore
+        run: dotnet restore
+        shell: bash
+
+      - name: Build Release (warnings as errors)
+        run: dotnet build --configuration Release --no-restore -warnaserror
+        shell: bash
