+++ .github/workflows/cd-develop-tst.yml
@@ -1,0 +1,98 @@
+name: Deploy develop to TST (CD)
+
+on:
+  push:
+    branches:
+      - develop
+    paths:
+      - 'src/**'
+      - 'docker/**'
+      - 'tests/**'
+      - '!docs/**'
+      - '!assets/**'
+      - '!team/**'
+  workflow_dispatch:
+
+concurrency:
+  group: cd-develop-tst
+  cancel-in-progress: false
+
+env:
+  REGISTRY: ghcr.io
+  IMAGE_NAME: workshopcode-api
+  SERVICE_NAME: seawind-api-tst
+
+jobs:
+  build-and-push-image:
+    runs-on: ubuntu-latest
+
+    permissions:
+      contents: read
+      packages: write
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Build-dotnet
+        uses: ./.github/actions/build-dotnet
+
+      - name: Run tests
+        run: dotnet test --filter "FullyQualifiedName!~Api.UnitTests.IntegrationTestDB" --configuration Release --no-build --no-restore --collect:"XPlat Code Coverage" --logger "trx"
+
+      - name: Generate HTML Coverage Report
+        uses: danielpalme/ReportGenerator-GitHub-Action@5
+        with:
+          reports: '**/coverage.cobertura.xml'
+          targetdir: 'coveragereport'
+          reporttypes: 'Html'
+
+      - name: Upload coverage report
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-html
+          path: coveragereport/
+
+      - name: Publish test results
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-results
+          path: |
+            **/TestResults/*.trx
+
+      - name: Publish Postgres log
+        if: always()
+        continue-on-error: true
+        uses: actions/upload-artifact@v4
+        with:
+          name: postgres-log
+          path: tests/Api.UnitTests/bin/Release/net8.0/postgres_log.txt
+
+      - name: Log in to container registry
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Extract metadata (tags, labels)
+        id: meta
+        uses: docker/metadata-action@v5
+        with:
+          images: ${{ env.REGISTRY }}/${{ github.repository_owner }}/${{ env.IMAGE_NAME }}
+          tags: |
+            type=sha
+            type=ref,event=branch
+
+      - name: Build and push Docker image
+        uses: docker/build-push-action@v6
+        with:
+          context: .
+          file: ./docker/Api.Dockerfile
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+
+      - name: Success log
+        run: echo "CD complete:Build + Docker push successful"
