+++ .github/workflows/ci-pr-develop.yml
@@ -1,0 +1,36 @@
+name: CI Pull Request (develop)
+
+on:
+  pull_request:
+    types: [opened, reopened, synchronize]
+    branches:
+      - develop
+
+jobs:
+  ci-pr-develop:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Build-dotnet
+        uses: ./.github/actions/build-dotnet
+
+      - name: Run tests
+        run: dotnet test --filter "FullyQualifiedName!~Api.UnitTests.IntegrationTestDB" --configuration Release --no-build --no-restore --logger "trx"
+
+      - name: Publish test results
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-results
+          path: |
+            **/TestResults/*.trx
+
+      - name: Publish Postgres log
+        if: always()
+        continue-on-error: true
+        uses: actions/upload-artifact@v4
+        with:
+          name: postgres-log
+          path: tests/Api.UnitTests/bin/Release/net8.0/postgres_log.txt
