+++ src/backend/Api/appsettings.IntegrationTests.json
@@ -1,4 +1,7 @@
 {
+  // Configuration overrides for integration tests ONLY.
+  // This file is NOT a full configuration and does NOT define a separate environment.
+  // It is applied conditionally via WC_USE_TEST_SETTINGS=true, on top of appsettings.json.
   "Seed": {
     "Preset": "Real",
     "MainDbContext": {
