+++ src/backend/Api/appsettings.json
@@ -5,5 +5,36 @@
       "Microsoft.AspNetCore": "Error"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "Seed": {
+    "Preset": "SuperFast", // "SeedOFF", "SeedCustom", "SuperFast", "Fast", "Real"
+    "CoreOption": {
+      "FileNameDataTemplate": "Demo{Entity}.json", // Пример для сущности Tester "DemoTester.json"
+      "FileNameSeedTemplate": "Demo{Entity}.seed.json", // Пример для сущности Tester "DemoTester.seed.json"
+      "SuffixSeederClassName": "Seeder" // Пример для сущности "Tester"+"Seeder" = "TesterSeeder"
+    },
+    "TimeDbContext": {
+      "Enabled": true,
+      "PathBase": "Time/SeedData", // Проект by "CopyToOutputDirectory" копирует файлы с данными из проекта в рабочий каталог
+      "Mode": "InsertOnly", // InsertOnly/InsertOrUpdate
+      //OPTION TURN OFF "PathVersion": "Time/SeedData/v1",
+      "UUID": "Real", // Stable/Real
+      "ExistenData": "NotDel", // Fresh/NotDel
+      "AutoMigrate": true // Возможно автоматической миграции в процессе сидирования
+    },
+    "MainDbContext": {
+      "Enabled": true,
+      "PathBase": "Main/SeedData", // Проект by "CopyToOutputDirectory" копирует файлы с данными из проекта в рабочий каталог
+      //OPTION TURN OFF "PathVersion": "Main/SeedData/v1",
+      "Mode": "InsertOnly", // InsertOnly/InsertOrUpdate
+      "UUID": "Stable", // Stable/Real
+      "ExistenData": "Fresh", // Fresh/NotDel
+      "AutoMigrate": true // Возможно автоматической миграции в процессе сидирования
+    }
+  },
+  "Jwt": {
+    "Issuer": "workshopcode",
+    "Audience": "workshopcode.api",
+    "Key": "TestJwtKeyForCIOnly_1234567890ABCD"
+  }
 }
