+++ tests/Api.UnitTests/Api.UnitTests.csproj
@@ -12,6 +12,7 @@
     <PackageReference Include="AutoFixture.AutoMoq" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
+	<PackageReference Include="coverlet.collector" />
     <PackageReference Include="Moq" />
 	<PackageReference Include="FluentAssertions" />
 	<PackageReference Include="FluentAssertions.Json" />
