+++ tests/Application.UnitTests/Application.UnitTests.csproj
@@ -10,6 +10,7 @@
 	<ItemGroup>
 		<PackageReference Include="AutoFixture.AutoMoq" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" />
+		<PackageReference Include="coverlet.collector" />
 		<PackageReference Include="Moq" />
 		<PackageReference Include="xunit" />
 		<PackageReference Include="xunit.runner.visualstudio">
